Refactor FilteredEnrichedTraceProvider into BufferedSourceFileFilter + UncoveredTraceProvider + UnionTraceProvider
